best_apr_backend/pyproject.toml
[tool.poetry]
name = "algebra-staking-backend"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
django-admin-rangefilter = "^0.8.2"
drf-spectacular = "^0.21.0"
django-filter = "^21.1"
django-debug-toolbar = "^3.2.2"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.1"
django-extensions = "^3.1.5"
web3 = "^5.25.0"
Werkzeug = "^2.0.2"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
celery = "^5.2.1"
django-redis = "^5.1.0"
django-celery-beat = "^2.2.1"
psycopg2-binary = "^2.9.2"
autopep8 = "^1.6.0"
gql = "^2.0.0"
jupyter = "^1.0.0"
ipython = "^7.30.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
